Fix TCP window update handling
Previously we would ignore window update packets since the ACK did not
acknowledge unacked data. This caused us to ignore window updates. This
is only really a problem when the remote side offers a zero-window. The
correct behavior is to periodically probe the remote side (as these
window-update packets could have been dropped anyways). We don't do this
yet.