fix(GeneralPurposeAllocator): fault handler for large regions
There was no page fault handler for large memory pages
in multi-coreand they were set to nullptrs, which cause
a secondary page fault. This fix keeps a table of virtual
to physical page mappings and uses it to do the same
mapping on cores that suffer the fault on a virtual address.

Pull request: #147
Approved by: dschatzberg