feat(startup): static ebb ClassInit
New protocol to allow static Ebbs to add logic
to the Init call provided by the template classes.
Specifically you must provide a ClassInit that will
be from within the provided Init call.  You may of
course provide an empty body.

Used this feature to fix the NodeAllocator initialization
bug.  NodeAllocator now provides a ClassInit that sets up
the global defaults.  This ensures that they are ready
when any call to the NodeAllocator is made including the
first call which may trigger a Ebb miss (which only then
creates the first rep).

Closes: #213